Update file structure for models
